backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
python-docx>=1.1.0
openai-whisper>=20231117
anthropic>=0.40.0
python-multipart>=0.0.9
sse-starlette>=2.1.0
pyannote.audio>=3.3.0
torch>=2.0.0
